Move SDC functions into aggregate module
These functions are used only in this module, so it makes sense to
locate them there. Doing so may also make an output-checker's life
easier.
